# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "API description",
        "contact": {
            "email": "contact@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all users",
                "description": "Get all users with pagination. __*Security: Richiede token M2M Passport*__",
                "operationId": "User.all",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Termine di ricerca (nome o email)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sortField",
                        "in": "query",
                        "description": "Campo per ordinamento",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sortOrder",
                        "in": "query",
                        "description": "Direzione (1 asc, -1 desc)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Elementi per pagina",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "create a new user",
                "description": "__*Security:*__ __*can be used only by clients with 'manager' role*__",
                "operationId": "User.create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "description": "User username",
                                        "type": "string",
                                        "example": "mario.rossi"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "description": "User password confirmation",
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "email": {
                                        "description": "User e-mail. It is not mandatory",
                                        "type": "string",
                                        "example": "mario.rossi@email.com"
                                    },
                                    "name": {
                                        "description": "User name",
                                        "type": "string",
                                        "example": "mario"
                                    },
                                    "surname": {
                                        "description": "User surname",
                                        "type": "string",
                                        "example": "rossi"
                                    },
                                    "enabled": {
                                        "description": "User enabled",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "password_expires_at": {
                                        "description": "User password expires at, format: Y-m-d H:i:s. If null, means the user is at his first login, so the password must be changed.",
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operation successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "OAuth2 Client Credentials Flow",
                "flows": {
                    "clientCredentials": {
                        "tokenUrl": "/oauth/token",
                        "scopes": {}
                    }
                }
            }
        }
    },
...
}